Actualizadas todas las clases para que funcionen con doubleWritable en vez de intWritable

+++ src/main/java/AsteroidDriver.java
@@ -1,5 +1,6 @@
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -30,7 +31,7 @@ public class AsteroidDriver extends Configured implements Tool {
         job.setReducerClass(AsteroidReducer.class);
 
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(DoubleWritable.class);
 
         return job.waitForCompletion(true) ? 0 : 1;
     }
