+++ src/main/java/AsteroidMapper.java
@@ -1,3 +1,4 @@
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
@@ -5,7 +6,7 @@ import org.apache.hadoop.mapreduce.Mapper;
 
 import java.io.IOException;
 
-public class AsteroidMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+public class AsteroidMapper extends Mapper<LongWritable, Text, Text, DoubleWritable> {
 
 
     @Override
@@ -18,12 +19,12 @@ public class AsteroidMapper extends Mapper<LongWritable, Text, Text, IntWritable
         String[] asteroid = line.split(",");
 
         String id = asteroid[0].replaceAll("\"", "");
-        int minYear = Integer.parseInt(asteroid[1].replaceAll("\"", "").split("-")[0]);
+        double minYear = Double.parseDouble(asteroid[1].replaceAll("\"", "").split("-")[0]);
 
         int diameter = Integer.parseInt(asteroid[6].replaceAll("\"", ""));
 
         if (minYear>=2021 && minYear <=2031){
-            context.write(new Text(id), new IntWritable(diameter));
+            context.write(new Text(id), new DoubleWritable(diameter));
         }
     }
 }
